Enhanced support for QTcpServer aside QLocalServer
In the original implementation by azadkuh backend type (either TCP or Local)
was managed by a member of type TBackend in most of classes. QHttpServer
does not allow to run TCP and Local server simultaneously and backend must
be selected at first but most of socket management methods you could find
an extra if-clause to determine which one of the inner methods must be called
based on the selected backend. I removed this extra If-clause by implementing
an abstract socket and determine its type right after creating server so count
of type checks and If-clauses has been decreased which causes higher throuput.